+++ viadot/flows/sql_server_to_duckdb.py
@@ -2,7 +2,7 @@ from prefect import Flow
 from typing import Any, Dict, List, Literal
 
 
-from ..task_utils import df_to_parquet, add_ingestion_metadata_task
+from ..task_utils import df_to_parquet, add_ingestion_metadata_task, map_dtypes_to_str
 from ..tasks import SQLServerToDF, DuckDBCreateTableFromParquet
 
 df_task = SQLServerToDF()
@@ -64,9 +64,9 @@ class SQLServerToDuckDB(Flow):
             config_key=self.sqlserver_config_key, query=self.sql_query, flow=self
         )
         df_with_metadata = add_ingestion_metadata_task.bind(df, flow=self)
-
+        df_mapped = map_dtypes_to_str.bind(df_with_metadata, flow=self)
         parquet = df_to_parquet.bind(
-            df=df_with_metadata,
+            df=df_mapped,
             path=self.local_file_path,
             if_exists=self.if_exists,
             flow=self,